Implement tests to respect user defined mapping and subsequent sync
